xdp_iphash_to_cpu_cmdline: fix usage of bpf_map_get_next_key
Fix the issue that looked like the xdp_iphash_to_cpu_cmdline --list
didn't list 0.0.0.0 entries and some entries were gone.

First change the variable name to be prev_key and key like:
  bpf_map_get_next_key(fd, prev_key, &key)

The prev_key need to be a real pointer, as we need to start with
prev_key = NULL.  Before prev_key was (a value on stack) implicit
zero. Thus, we basically asked to start the listing with key 0.
Thus, it is why it *looked* like it was all working when the hash
doesn't contain key=0, as then it fails-to-start-at-0 and kernel
provides the first-key.

Signed-off-by: Jesper Dangaard Brouer <netoptimizer@brouer.com>