bin/tc_mq_htb_setup_example.sh: Numeric sort TX queue dirs
Like xps_setup.sh numerically sorted TX queues.

This change is not really needed, as the $dir variable is not really
used for anything.  The code only need a counter for number of TX-queues.

Do change anyway as this is an example script that we expect people
to modify for their own usage.  Hopefully this highlight the pitfall
with the lexical sorted TX queues.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>