Convert IPs to IPv4-mapped IPv6 addresses in preparation for v6 support
This converts the IP address representation to use IPv4-mapped IPv6
addresses for all v4 addresses. The only functional change should be that
the output of --list now outputs these v4-mapped addresses.

This also makes it possible to insert real IPv6 addresses into the config
from userspace, but there is no support for parsing v6 packets in the
kernel BPF programs yet. This will be added in a subsequent commit.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>