shared_maps.h, common_kern_user.h: Change map_ip_hash to a BPF_MAP_TYPE_LPM_TRIE
It's very common in IPv4 networks to need to apply a shaper rule to a subnet, rather than an
individual host. For example, if a customer occupies the entire 192.168.100.0/24 subnet - but
the provider is selling a single 100 Mbps connection - the entire subnet will likely fall
inside a single classifier rule. Previous versions of xdp-cpumap-tc would require the creation
of 254 entries in map_ip_hash to accomplish this - quickly exhausting the 32k rules limit for
a large Internet Service Provider.

The problem becomes more extreme when IPv6 is employed. If RIPE's recommendations [1] are
applied, each customer receives a /64 allocation. Creating 2^64 map entries per customer
is impractical (and won't work at all with IP_HASH_ENTRIES set to 32,767).

In 2017, BPF_MAP_TYPE_LPM_TRIE was added to the Linux BPF implementation to avoid this problem.[2]
BPF_MAP_TYPE_LPM_TRIE implements a "longest prefix match" unbalanced Trie: keys are stored as a
combined "prefix length" and address. For example, 192.168.100.0/24 would be stored as:

* prefix: 24
* address: 192.168.100.0

Longest Prefix Matching (LPM) ensures that the most-specific match is selected, falling back to
larger subnets/smaller prefixes when a specific match is not found. For example, 192.168.100.5 would
match 192.168.100.0/24 in the absence of a more-specific key.

This commit creates a new type, named ip_hash_key, to replace the __u32 specific addresses
used in previous versions. The key contains a prefix (which must be in the range 0 through 128)
and 128-bits of address, in the form of an IPv6 address. IPv4 addresses should be stored in the
last 32-bits, with the preceeding bits zeroed. Queries should be performed with a starting prefix
of 128, ensuring the most-specific IPv4 or IPv6 match is returned.

The new map style reduces storage requirements for an IPv4 /24 block from 254 entries to
1. Likewise, an IPv6 /64 requires 1 map entry rather than 18,446,744,073,709,551,616.

If you are classifying subnets, this represents a significant improvement in memory utilization.
If you are only classifying individual (/32) IPv4 addresses, then the new approach adds overhead:
each address now consumes 160 bits (128-bit address, 32-bit prefix) rather than 32.

Using a Trie as opposed to a map also imposes a small performance penalty when compared with
individual address lookups. Directly hashing a 32-bit address (into a 32-bit lookup value)
is faster than tree-walking. This is a trade-off for the flexibility of being able to classify
an entire subnet (or IPv6 address) with a single map entry.

[1] - https://www.ripe.net/publications/docs/ripe-690
[2] - https://github.com/torvalds/linux/commit/b95a5c4db09bc7c253636cb84dc9b12c577fd5a0

Signed-off-by: Herbert Wolverson <herberticus@gmail.com>