xdp_iphash_to_cpu_kern.c: Replace parse_ipv4 with parse_ip
Replaces the parse_ipv4 function with a new parse_ip function (derived
from a similar function in the IPv6 branch of this repository). The
new function reads the eth_proto of each packet, and extracts IPv4 or
IPv6 addresses into an ip_hash_key structure. IPv4 addresses are packed
into the last 32-bits of the structure's address (with the preceding
items zeroed), IPv6 addresses consume the entire lookup address.

Once the address has been parsed, a lookup (starting with a prefix of
128, using longest prefix matching to ensure a most-specific match) is
used to return the mapping information for an IP address. The program
then continues as it did before - assigning a CPU to the incoming
packet.

Signed-off-by: Herbert Wolverson <herberticus@gmail.com>