common_user.h/common_user.c: Update IP address parsing.
Adds a helper function (ip_string_to_key) to convert IP address strings
to ip_hash_key structures. The function uses getaddrinfo to parse
user input and determine whether an address was an IPv4, IPv6 or invalid
address.

* IPv4 addresses are packed into the last 32-bits of the lookup address,
  with the preceding entries zeroed.
* IPv6 address occupy the key's entire address field.
* Invalid addresses return a prefix-length of 0 to indicate invalid.

Prefix length is stored as either the provided prefix + 96 for IPv4 (to
account for the 128-bit storage scheme) or as specified for IPv6.

Adds a "prefix" field to the iphash_modify function, which is passed to
the ip_string_to_key function. iphash_modify is adjusted to return
IP_FAIL if ip_string_to_key returns a prefix length of 0, indicating
an invalid address.

Signed-off-by: Herbert Wolverson <herberticus@gmail.com>