common_user.c, common_user.h, xdp_iphash_to_cpu_cmdline.c: Add direct parsing of CIDR addresses
Updates the ip_string_to_key function to search the provided IP address string
for a '/'. If one is provided, the prefix length is extracted from the subsequent
digits. If no prefix is provided, default to etiher /32 (IPv4) or /128 (IPv6).

Removes the requirement for a prefix parameter for ip_string_to_key and
ip_hash_modify, and restores the functions' original headers. Adjusts the
calls to these functions in xdp_ip_hash_to_cpu_cmdline.c.

Removes the --prefix CLI option, since it is no longer needed.

Signed-off-by: Herbert Wolverson <herberticus@gmail.com>