common_user.c: Replace 0 with 255 as the error-code for failing to parse an IP address.
ip_string_to_key indicates success or failure with the prefixlen of the
returned ip_hash_key. Previously, it returned 0 to indicate failure.
This would collide with deliberately using a prefix of 0 (for a default
match), preventing 0.0.0.0/0 from being added to the lookup map.

The error code is now 255, which should never occur for an IPv4 or
IPv6 address.

Signed-off-by: Herbert Wolverson <herberticus@gmail.com>