Change IPv4 packing to a 0xFFFFFFFF prefix from 0x0.
Previous versions packed IPv4 keys as 96 bits of zeroes, followed by the
32-bit IPv4 address. While this worked, it caused some issues:

* Specifying 0.0.0.0/0 as a default handler resulted in the lister
  deducing that the address was an IPv4 address and displaying
  the incorrect value.
* Current users relied on matching 255.255.255.255 for default
  queue handling. This was broken in this patch.

This commit replaces the zero padding with all ones. So an
ip_hash_key containing an IPv4 address is now represented as:
   0xFFFFFF 0xFFFFFFFF 0xFFFFFFFF (the address)

All default setup of ip_hash_key structures are replaced with
the new scheme in this patch.

Handling of 255.255.255.255 as a default is re-enabled in
both xdp_iphash_to_cpu_kern.c and tc_classify_kern.c. This
imposes a small overhead (a second map lookup), but preserves
compatibility.

A preferred approach is to specify "::/0" as a default address
(an IPv6 default route). This will match any provided address/
prefix combination, and not incur the cost of a secondary route
lookup.

Tested with:

* 0.0.0.0/0 as a default, which correctly matches IPv4 as a default.
  It does NOT match unspecified IPv6 addresses. You can use this
  if you desire alternative behavior for un-matched IPv4 vs IPv6.
* ::/0 as a default, correctly matched both IPv4 and IPv6. You
  can use this to match any address.
* 255.255.255.255 (/32 - not specified), which correctly applied the
  default to IPv4 addresses. Will not trigger a match for IPv6
  addresses.

Signed-off-by: Herbert Wolverson <herberticus@gmail.com>